LPS-176935 Split SAPEntry creation to its own PortalInstanceLifecycleListener and marked as persistent
