+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/instance/lifecycle/SAPEntryPortalInstanceLifecycleListener.java
@@ -1,0 +1,81 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.object.internal.instance.lifecycle;
+
+import com.liferay.object.constants.ObjectSAPConstants;
+import com.liferay.portal.instance.lifecycle.BasePortalInstanceLifecycleListener;
+import com.liferay.portal.instance.lifecycle.PortalInstanceLifecycleListener;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.model.Company;
+import com.liferay.portal.kernel.service.ServiceContext;
+import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.util.ResourceBundleUtil;
+import com.liferay.portal.language.LanguageResources;
+import com.liferay.portal.security.service.access.policy.model.SAPEntry;
+import com.liferay.portal.security.service.access.policy.service.SAPEntryLocalService;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Shuyang Zhou
+ */
+@Component(service = PortalInstanceLifecycleListener.class)
+public class SAPEntryPortalInstanceLifecycleListener
+	extends BasePortalInstanceLifecycleListener {
+
+	@Override
+	public void portalInstanceRegistered(Company company) {
+		try {
+			_addSAPEntry(company.getCompanyId());
+		}
+		catch (PortalException portalException) {
+			_log.error(
+				"Unable to add service access policy entry for company " +
+					company.getCompanyId(),
+				portalException);
+		}
+	}
+
+	private void _addSAPEntry(long companyId) throws PortalException {
+		SAPEntry sapEntry = _sapEntryLocalService.fetchSAPEntry(
+			companyId, ObjectSAPConstants.SAP_ENTRY_NAME);
+
+		if (sapEntry != null) {
+			return;
+		}
+
+		_sapEntryLocalService.addSAPEntry(
+			_userLocalService.getDefaultUserId(companyId),
+			ObjectSAPConstants.ALLOWED_SERVICE_SIGNATURES, true, true,
+			ObjectSAPConstants.SAP_ENTRY_NAME,
+			ResourceBundleUtil.getLocalizationMap(
+				LanguageResources.PORTAL_RESOURCE_BUNDLE_LOADER,
+				"service-access-policy-entry-default-object-title"),
+			new ServiceContext());
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		SAPEntryPortalInstanceLifecycleListener.class);
+
+	@Reference
+	private SAPEntryLocalService _sapEntryLocalService;
+
+	@Reference
+	private UserLocalService _userLocalService;
+
+}
